+++ src/main/java/joe/Parser.java
@@ -1,8 +1,12 @@
+package joe;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
+import joe.commands.*;
+import joe.exceptions.JoeException;
 
 public class Parser {
   private enum CommandType {
@@ -145,7 +149,8 @@ public class Parser {
     }
     sb.setLength(sb.length() - 2);
     String msg =
-        String.format("Invalid Command Keyword!%nHere is a list of valid commands: %s", sb);
+        String.format(
+            "Invalid Command Keyword!%nHere is a list of valid commands: %s", sb);
     return new InvalidCommand(msg);
   }
