+++ src/main/java/joe/Ui.java
@@ -1,3 +1,5 @@
+package joe;
+
 import java.util.Scanner;
 
 public class Ui {
