+++ src/main/java/joe/commands/Command.java
@@ -1,0 +1,17 @@
+package joe.commands;
+
+import joe.Storage;
+import joe.TaskList;
+import joe.Ui;
+import joe.exceptions.JoeIndexOutOfBoundsException;
+
+public abstract class Command {
+  protected boolean isExit = false;
+
+  public abstract void execute(TaskList tasks, Ui ui, Storage storage)
+      throws JoeIndexOutOfBoundsException;
+
+  public boolean isExit() {
+    return isExit;
+  }
+}
