+++ src/main/java/joe/commands/ListCommand.java
@@ -1,0 +1,13 @@
+package joe.commands;
+
+import joe.Storage;
+import joe.TaskList;
+import joe.Ui;
+
+public class ListCommand extends Command {
+
+  @Override
+  public void execute(TaskList tasks, Ui ui, Storage storage) {
+    ui.print(tasks.toString());
+  }
+}
