+++ src/main/java/joe/exceptions/JoeException.java
@@ -1,3 +1,5 @@
+package joe.exceptions;
+
 public class JoeException extends Exception {
   public JoeException(String error) {
     super(error);
