+++ src/main/java/joe/exceptions/JoeIndexOutOfBoundsException.java
@@ -1,3 +1,5 @@
+package joe.exceptions;
+
 public class JoeIndexOutOfBoundsException extends JoeException {
   public JoeIndexOutOfBoundsException(int idx) {
     super("Task " + idx + " does not exist");
