+++ src/main/java/joe/tasks/Task.java
@@ -1,3 +1,5 @@
+package joe.tasks;
+
 public class Task {
   protected String description;
   protected boolean isDone;
