+++ text-ui-test/runtest.bat
@@ -10,7 +10,7 @@ REM delete saved task list from previous run
 if exist joe.TXT del joe.TXT
 
 REM compile the code into the bin folder
-javac  -cp ..\src\main\java -Xlint:none -d ..\bin ..\src\main\java\*.java
+javac  -cp ..\src\main\java\ -Xlint:none -d ..\bin ..\src\main\java\joe\*.java ..\src\main\java\joe\commands\*.java ..\src\main\java\joe\exceptions\*.java ..\src\main\java\joe\tasks\*.java
 IF ERRORLEVEL 1 (
     echo ********** BUILD FAILURE **********
     exit /b 1
